fix(dashboard): return 404 for missing deployments or projects
Replace runtime errors thrown when a deployment or project is not found
with Next.js notFound(), rendering a 404 boundary instead of a 500.
