+++ web/apps/dashboard/app/(app)/[workspaceSlug]/projects/[projectId]/not-found.tsx
@@ -1,0 +1,8 @@
+"use client";
+import { NotFoundState } from "@/components/not-found-state";
+
+export default function ProjectNotFound() {
+  return (
+    <NotFoundState description="We couldn't find the project or deployment that you're looking for." />
+  );
+}
